requirements.txt
flask
flask_sqlalchemy
psycopg2
python-dotenv
gunicorn

#testing
coverage
pytest
pytest-dotenv